Timing: gameplay time is double precision, spatial simulation stays float
A Hard Dragon breathed sixteen fireballs per breath where the Python
source breathes fifteen.  That was not a Dragon bug and it has not been
fixed as one.

`1f/60f` is 0.016666668 -- fractionally LONGER than a true sixtieth --
so a float timer that subtracts it drifts a little every step, and at a
comparison boundary that becomes a whole extra event.  Measured across
the six fixtured breath configurations, three diverged, and not all in
the same direction:

    fireScale  breath  interval   Python   float32
       1.0      0.50     0.15        4        3      -1
       0.5      1.25     0.15       15       16      +1   <- shipped Hard
       0.5      0.50     0.15        7        6      -1

A divergence that goes both ways cannot be compensated for by adjusting
a constant.  The type was wrong.

  Gameplay time  = double     durations, deadlines, cooldowns, intervals,
                              elapsed and remaining time, schedules
  Spatial maths  = float      x/y, velocities, dimensions, angles,
                              collision geometry, drawing state

WHAT CHANGED

`Simulation.Stepper.step` now takes a `double` and is handed FIXED_DT
itself, all the way down through `GameWorld.step`, `StepListener.onStep`
and every `update`/`think`.  A system that integrates space narrows the
step once, itself, at the top of the method:

    public void update(double dt) {
        float fdt = (float) dt;     // spatial only
        cooldown -= dt;             // time domain: never fdt
        x += vx * fdt;              // spatial domain
    }

`Simulation.DT` is renamed `PHYSICS_DT` so that a call site reading one
is visibly a spatial call site.  No production class reads it any more;
a test fails the build if one starts to.  There is still exactly ONE
clock -- FIXED_DT and PHYSICS_DT are the same step at two precisions,
and nothing advances them independently.

45 timer fields converted across defences, enemies, bosses and the
cursor, plus the durations they are loaded from: 10 GameConfig
constants, 13 in Tuning, and the duration fields of EnemyConfig,
BossConfig, TowerConfig and DifficultyConfig.  `Json5.seconds` and
`Rng.uniformSeconds` were added so a duration is visibly a duration
where it is read or drawn; uniformSeconds consumes exactly one nextLong
like uniform, so the gameplay stream advances identically and a seeded
run still replays.

WHAT DELIBERATELY DID NOT CHANGE

  * Not integer ticks.  Durations stay seconds -- many are random or
    configurable fractions -- they are simply double seconds.
  * Not epsilons.  Every source comparison operator is preserved; `<= 0`
    stays `<= 0`.  No EPSILON was introduced to make a test green.
  * Not a global conversion.  Rates that are not times stay float:
    `regen` is HP/s, `breathPower` a damage fraction, `shove` a
    velocity.  So does drawing state: hurtFlash, recoil, aura, trapGlow,
    orb, smash, fuse, anim, bob, spin.
  * No gameplay constant was adjusted.

RESULTS

  * Dragon Hard breath: Python 15, Java 15.  Asserted in the standalone
    cadence loop AND against a live Dragon, exactly -- no tolerance,
    because the boundary is no longer precision-dependent.
  * Ballista/Cannon reloads are now the values castle.py states.  The
    old assertions expected 3.0999999046325684; a reload of 3.1 s is now
    3.1 s.  Same for the boss constants: REGALIA_CD_GROWTH is 0.6, not
    0.6000000238418579.
  * The float32 fixtures are KEPT, as a regression demonstration of why
    the rule exists rather than as a description of production.
  * 505 tests, 0 failures.  Was 495.

The Python sources remain untouched and authoritative.

+++ java-port/core/src/main/java/com/mymmer/castledefense/boss/Boss.java
@@ -42,7 +42,7 @@ public abstract class Boss extends Enemy {
     protected final BossContext bossCtx;
 
     /** Seconds of arrival animation. Visual state; nothing gates on it yet. */
-    private float intro;
+    private double intro;
     /** Visual only: the hit glow, decaying at 3/s. */
     private float aura;
 
@@ -53,10 +53,10 @@ public abstract class Boss extends Enemy {
      * Captured once at construction, so a difficulty change mid-run cannot
      * retune a boss already on the field.
      */
-    private final float fireScale;
+    private final double fireScale;
 
     /** Seconds until the regalia can be disrupted again. */
-    private float regaliaCd;
+    private double regaliaCd;
     /** How many times this boss has been disrupted. Drives the guard growth. */
     private int regaliaTaken;
 
@@ -87,7 +87,7 @@ public abstract class Boss extends Enemy {
     }
 
     /** Visual only. */
-    public float intro() {
+    public double intro() {
         return intro;
     }
 
@@ -96,18 +96,25 @@ public abstract class Boss extends Enemy {
         return aura;
     }
 
-    /** Scales a reload time by the difficulty's boss fire rate. */
-    public float fireDelay(float seconds) {
+    /**
+     * Scales a reload time by the difficulty's boss fire rate.
+     *
+     * <p>Time in, time out: {@code double} at both ends. Hard's 0.5 turns the
+     * Dragon's 0.15 s shot interval into 0.075 s, and doing that in floats is
+     * precisely what made a Hard breath fire sixteen fireballs where Python
+     * fires fifteen.
+     */
+    public double fireDelay(double seconds) {
         return seconds * fireScale;
     }
 
-    public float fireScale() {
+    public double fireScale() {
         return fireScale;
     }
 
     // --- the disruption guard -----------------------------------------------
 
-    public float regaliaCd() {
+    public double regaliaCd() {
         return regaliaCd;
     }
 
@@ -126,14 +133,14 @@ public abstract class Boss extends Enemy {
      */
     public void guardRegalia() {
         regaliaCd = GameConfig.REGALIA_COOLDOWN
-                * (1f + GameConfig.REGALIA_CD_GROWTH * regaliaTaken);
+                * (1d + GameConfig.REGALIA_CD_GROWTH * regaliaTaken);
     }
 
     /** Counts one successful disruption. Call <em>before</em> {@link #guardRegalia}. */
     protected void countDisruption() {
         regaliaTaken++;
         bossCtx.trace().event(TraceEvent.BOSS_DISRUPTION, bossCtx.step(), uid(),
-                regaliaTaken, regaliaCd, boss.type.id());
+                regaliaTaken, (float) regaliaCd, boss.type.id());
     }
 
     /**
@@ -163,7 +170,7 @@ public abstract class Boss extends Enemy {
      * subclass says its item is not there to be taken.
      */
     public boolean regaliaCovers(float px, float py) {
-        if (!isAlive() || regaliaCd > 0f || !hasRegalia()) {
+        if (!isAlive() || regaliaCd > 0d || !hasRegalia()) {
             return false;
         }
         if (!regaliaAnchorInto(anchorOut)) {
@@ -219,10 +226,10 @@ public abstract class Boss extends Enemy {
     }
 
     @Override
-    public void update(float dt) {
-        aura = Math.max(0f, aura - dt * 3f);
-        intro = Math.max(0f, intro - dt);
-        regaliaCd = Math.max(0f, regaliaCd - dt);
+    public void update(double dt) {
+        aura = Math.max(0f, aura - (float) dt * 3f);       // visual
+        intro = Math.max(0d, intro - dt);
+        regaliaCd = Math.max(0d, regaliaCd - dt);
         super.update(dt);
     }
